Fix tests, move AuthorityAliasesTests
